requirements.txt
langchain_openai
langchain_core
langchain_community
langserve
python_dotenv
streamlit
